Methods for objects of class "flite"

Methods for objects of class "flite" returned from flite.

Usage

# S3 method for class 'flite'
plot(
  x,
  which = c("all", "pu", "gp", "xi", "theta"),
  adj_type = c("vertical", "none", "cholesky", "spectral"),
  ...
)

# S3 method for class 'flite'
coef(object, ...)

# S3 method for class 'flite'
vcov(object, adjust = TRUE, ...)

# S3 method for class 'flite'
nobs(object, ...)

# S3 method for class 'flite'
logLik(object, ...)

# S3 method for class 'flite'
summary(object, adjust = TRUE, digits = max(3, getOption("digits") - 3L), ...)

# S3 method for class 'summary.flite'
print(x, ...)

# S3 method for class 'flite'
confint(
  object,
  parm = "all",
  level = 0.95,
  adj_type = c("vertical", "none", "cholesky", "spectral"),
  profile = TRUE,
  ...
)

Arguments

x

An object inheriting from class "flite", a result of a call to flite.

which

A character scalar indicating which plot(s) to produce. If which = "all" then all 4 plots described in Details are produced. Otherwise, only one of these plots is produced, with the possible names of the arguments being in the order that the plots are described in Details.

adj_type

A character scalar passed to conf_intervals and conf_region as the argument type to select the type of adjustment applied to the independence log-likelihood. Of the 3 adjustments, "vertical" is preferred because it preserves constraints on the parameters, whereas the "cholesky" and "spectral" adjustment do not. In the generalised Pareto case the constraint that \(\xi\) > -\(\sigma\)u / \(x\)(n) where \(x\)(n) is the largest excesses of the threshold \(u\), is preserved.

...

For plot.flite: arguments passed to plot, such as graphical parameters.

For print.summary.flite: additional arguments passed to print.default.

For confint.flite: additional arguments passed to conf_intervals.

Otherwise ... is unused.

object

An object of class "flite", returned by flite.

adjust

A logical scalar. If adjust = TRUE then the elements of the variance-covariance matrix corresponding to (\(p\)u, \(\sigma\)u, \(\xi\)), are estimated using a sandwich estimator. See flite. Otherwise, this matrix is the inverse of the observed information matrix.

digits

An integer. Passed to signif to round the values in the summary.

parm

A character vector specifying the parameters for which confidence intervals are required. The default, which = "all", produces confidence intervals for all the parameters, that is, \(p\)u, \(\sigma\)u, \(\xi\) and \(\theta\). If which = "gp" then intervals are produced only for \(\sigma\)u and \(\xi\). Otherwise, parm must be a subset of c("pu", "sigmau", "xi", "theta").

level

The confidence level required. A numeric scalar in (0, 1).

profile

A logical scalar. If TRUE then confidence intervals based on an (adjusted) profile loglikelihood are returned. If FALSE then intervals based on approximate large sample normal theory, which are symmetric about the MLE, are returned.

Value

plot.flite: No return value, only the plot is produced.

coef.flite: a numeric vector of length 4 with names c("p[u]", "sigma[u]", "xi", "theta"). The MLEs of the parameters \(p\)u, \(\sigma\)u, \(\xi\) and \(\theta\).

vcov.flite: a \(4 \times 4\) matrix with row and column names c("p[u]", "sigma[u]", "xi", "theta"). The estimated variance-covariance matrix for the model parameters. If adjust = TRUE then the elements corresponding to \(p\)u, \(\sigma\)u, and \(\xi\) are adjusted for cluster dependence using a sandwich estimator; otherwise they are not adjusted.

nobs.flite: a numeric vector of length 3 with names c("p[u]", "gp", "theta"). The respective number of observations used to estimate \(p\)u, (\(\sigma\)u, \(\xi\)) and \(\theta\).

logLik.flite: an object of class "logLik": a numeric scalar with value equal to the maximised log-likelihood. This is the sum of contributions from three fitted models, from a Bernoulli model for occurrences of threshold exceedances, a generalised Pareto model for threshold excesses and a \(K\)-gaps model for the extremal index. The returned object also has attributes nobs, the numbers of observations used in each of these model fits, and "df" (degrees of freedom), which is equal to the number of total number of parameters estimated (4).

summary.flite: an object containing the original function call and a matrix of estimates and estimated standard errors with row names c("p[u]", "sigma[u]", "xi", "theta"). The object is printed by print.summary.flite.

print.summary.flite: the argument x is returned, invisibly.

confint.flite: a numeric matrix with 2 columns giving the lower and upper confidence limits for each parameter. These columns are labelled as (1-level)/2 and 1-(1-level)/2, expressed as a percentage, by default 2.5% and 97.5%. The row names are the names of the parameters supplied in parm.

Details

For plot.flite, if which = "all" then 4 plots are produced.

Top left: (adjusted) log-likelihood for the threshold exceedence probability \(p\)u, with a horizontal line indicating a 95% confidence interval for \(p\)u.

Top right: contour plot of the (adjusted) log-likelihood for the GP parameters (\(\sigma\)u, \(\xi\)), showing (25, 50, 75, 90, 95)% confidence regions. The linear constraint \(\xi\) > -\(\sigma\)u / \(x\) (n) is drawn on the plot.

Bottom left: (adjusted) log-likelihood for \(\xi\), with a horizontal line indicating a 95% confidence interval for \(\xi\).

Bottom right: log-likelihood for the extremal index \(\theta\), with a horizontal line indicating a 95% confidence interval for \(\theta\).

See also