Methods for objects of class "blite"

Methods for objects of class "blite" returned from blite. confint.blite is a misnomer: it returns (equi-tailed) Bayesian credible intervals.

# S3 method for blite
plot(x, which = c("all", "pu", "gp", "xi", "theta"), ...)

# S3 method for blite
coef(object, fun, ...)

# S3 method for blite
vcov(object, ...)

# S3 method for blite
nobs(object, ...)

# S3 method for blite
summary(
  object,
  short = TRUE,
  mean = TRUE,
  digits = max(3, getOption("digits") - 3L),
  ...
)

# S3 method for summary.blite
print(x, ...)

# S3 method for blite
confint(object, parm = "all", level = 0.95, ...)

Arguments

x

An object inheriting from class "blite", a result of a call to blite.

which

A character scalar indicating which plot(s) to produce. If which = "all" then all 4 plots described in Details are produced. Otherwise, only one of these plots is produced, with the possible names of the arguments being in the order that the plots are described in Details.

...

For plot.blite: arguments passed to plot, such as graphical parameters.

For coef.blite: additional arguments passed to fun.

For print.summary.blite: additional arguments passed to print.default.

Otherwise ... is unused.

object

An object of class "blite", returned by blite.

fun

A summary function to be applied to each column of the simulated values in object. If fun is missing then mean is used.

short

A logical scalar that determines the form of the output. See Details.

mean

A logical scalar. Determines the form of the output if short = TRUE. See Details.

digits

An integer. Passed to signif to round the values in the summary.

parm

A character vector specifying the parameters for which confidence intervals are required. The default, which = "all", produces confidence intervals for all the parameters, that is, \(p\)u, \(\sigma\)u, \(\xi\) and \(\theta\). If which = "gp" then intervals are produced only for \(\sigma\)u and \(\xi\). Otherwise, parm must be a subset of c("pu", "sigmau", "xi", "theta").

level

The credible level required. A numeric scalar in (0, 1).

Value

plot.blite: No return value, only the plot is produced.

coef.blite: a numeric vector of length 4 with names

c("p[u]", "sigma[u]", "xi", "theta"). The values of summary statistics calculated using the function fun.

vcov.blite: a \(4 \times 4\) matrix with row and column names c("p[u]", "sigma[u]", "xi", "theta"). An estimate of the posterior covariance matrix, calculated using

cov.

nobs.blite: a numeric vector of length 3 with names

c("p[u]", "gp", "theta"). The respective number of observations used to infer \(p\)u, (\(\sigma\)u,

\(\xi\)) and \(\theta\).

summary.blite: an object containing the original function call and a matrix of summaries of the posterior samples for each of the parameters. If short = TRUE then there are 2 columns, containing either the sample posterior means and standard deviations (mean = TRUE) or the sample posterior medians and inter-quartile ranges (mean = FALSE). If short = FALSE then there are 4 columns, with each column containing the usual 6-number summary produced by summary. The object is printed by

print.summary.blite.

print.summary.blite: the argument x is returned, invisibly.

confint.blite: a numeric matrix with 2 columns giving the lower and upper credible limits for each parameter. These columns are labelled as (1-level)/2 and 1-(1-level)/2, expressed as a percentage, by default 2.5% and 97.5%. The row names are the names of the parameters supplied in parm.

Details

For plot.blite, if which = "all" then 4 plots are produced.

Top left: histogram of the posterior sample for the threshold exceedance probability \(p\)u.

Top right: scatter plot of posterior sample for the GP parameters (\(\sigma\)u, \(\xi\)). The linear constraint \(\xi\) > -\(\sigma\)u / \(x\) (n) is drawn on the plot.

Bottom left: histogram of the posterior sample for the GP shape parameter \(\xi\).

Bottom right: histogram of the posterior sample for the extremal index \(\theta\).

See also