Methods for objects of class "returnLevel"

Methods for objects of class "returnLevel" returned from returnLevel.

# S3 method for returnLevel
plot(x, level = NULL, legend = TRUE, digits = 3, plot = TRUE, ...)

# S3 method for returnLevel
print(x, digits = max(3L, getOption("digits") - 3L), ...)

# S3 method for returnLevel
summary(object, digits, ...)

# S3 method for summary.returnLevel
print(x, ...)

Arguments

x

an object of class c("returnLevel", "lite"), a result of a call to returnLevel, using prof = TRUE.

level

A numeric scalar in (0, 1). The confidence level required for the confidence interval for the m-year return level. If level is not supplied then x$level is used. level must be no larger than x$level.

legend

A logical scalar. Should we add a legend (in the top right of the plot) that gives the approximate values of the MLE and 100level% confidence limits?

digits

For plot.returnLevel: an integer. Passed to signif to round the values in the legend.

For print.returnLevel: the argument digits to print.default.

For summary.returnLevel: an integer. For number formatting with signif. If digits is not specified (i.e. missing) then signif() will not be called (i.e. no rounding will be performed).

plot

A logical scalar. If TRUE then the plot is produced. Otherwise, it is not, but the MLE and confidence limits are returned.

...

For plot.returnLevel: arguments passed to plot, such as graphical parameters.

For print.summary.returnLevel: arguments passed to print.default.

object

an object of class c("returnLevel", "lite"), a result of a call to returnLevel, using prof = TRUE.

Value

plot.returnLevel: a numeric vector of length 3 containing the lower 100level% confidence limit, the MLE and the upper 100level% confidence limit is returned invisibly. print.returnLevel: the argument x is returned, invisibly. summary.returnLevel: a list containing the list element object$call and a matrix matrix containing the MLE and estimated SE of the return level. print.summary.returnLevel: the argument x is returned, invisibly.

Details

plot.returnLevel plots the profile log-likelihood for a return level, provided that x returned by a call to returnLevel using prof = TRUE. Horizontal lines indicate the values of the maximised log-likelihood and the critical level used to calculate the confidence limits. If level is smaller than x$level then approximate 100level% confidence limits are recalculated based on the information contained in x$for_plot.

print.returnLevel prints the call to returnLevel and the estimates and 100x$level% confidence limits for the x$m-year return level.

Examples

See returnLevel.

See also